Add pluggable agents framework
This commit introduces the --agent option, a string argument that tells the simulator to look for a vacuum agent with the given name in the module load path and evaluate it.

Because the simulator can now be run with third-party agents, the experiment has been made more robust to agent crashes and illegal decisions.