Convert basic environment to new config API
This commit introduces a pluggable architecture for environments exactly like the one for agents, and introduces a new API for passing arguments from the command line into plugged-in environments.

Any class that implements the update(self, action) method and the state and observable_state properties can be used as an environment. To specify an environment on the command line, use '--environment IMPORT_PATH', where IMPORT_PATH is any importable module and class name, like 'my_environment.MyEnvironment'.

Any initialization parameters may be passed into the environment, by specifying '--env-PARAMNAME [PARAMVALUE ...]'. The prefix is stripped from the flag and the values are collated into a list of strings, and the parameter is passed as a keyword argument to the environment class's __init__ method as PARAMNAME=[PARAMVALUE, ...].

It is up to the user to specify and environments that interoperate with each other. The BasicVacuumWorld environment has been modified to implement this interface.