Implement RoombaWorld travelling salesman problem
This commit adds a fully implemented problem formulation for the Travelling Salesman Problem in a RoombaWorld with known initial geography. The initial state is set from an agent starting location and an iterable of passable locations, and initializes a set of visited locations with the agent's current location. The action space and transition model permit moving one space in either the X or Y direction, provided the new location is in the set of passable locations. The goal test checks for equality of the visited locations list and the list of all passable locations. The path cost is constant.